modify hardware_id(tinyint->int), inventory_qty(tinyint->int), alert_qty(tinyint->int) then migrate

+++ database/migrations/2022_11_21_142132_create_inventories_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
     {
         Schema::create('inventories', function (Blueprint $table) {
             $table->id();
-            $table->tinyInteger('hardware_id');
-            $table->tinyInteger('inventory_qty');
-            $table->tinyInteger('alert_qty');
+            $table->integer('hardware_id');
+            $table->integer('inventory_qty');
+            $table->integer('alert_qty');
             $table->timestamps();
         });
     }
